# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "Transaction API for client transactions",
    "title": "Transaction API",
    "version": "1.0.0"
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "paths": {
...
    "/api/v1/users/{userId}/category-overrides": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "category": {
                            "type": "string"
                          },
                          "id": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          },
                          "label": {
                            "type": "string"
                          },
                          "mappedTo": {
                            "additionalProperties": false,
                            "properties": {
                              "category": {
                                "type": "string"
                              },
                              "id": {
                                "maximum": 9007199254740991,
                                "minimum": -9007199254740991,
                                "type": "integer"
                              },
                              "label": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "category",
                              "label",
                              "updatedAt"
                            ],
                            "type": "object"
                          }
                        },
                        "required": [
                          "id",
                          "category",
                          "label"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "links": {
                      "additionalProperties": false,
                      "properties": {
                        "next": {
                          "nullable": true,
                          "type": "string"
                        },
                        "prev": {
                          "nullable": true,
                          "type": "string"
                        },
                        "self": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "self",
                        "next",
                        "prev"
                      ],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": {
                        "count": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "count"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data",
                    "links",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user categories"
        ]
      }
    },
    "/api/v1/users/{userId}/category-overrides/{categoryId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "exclusiveMinimum": true,
              "maximum": 9007199254740991,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user categories"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "exclusiveMinimum": true,
              "maximum": 9007199254740991,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "toCategoryId": {
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "toCategoryId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "category": {
                      "type": "string"
                    },
                    "id": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "label": {
                      "type": "string"
                    },
                    "mappedTo": {
                      "additionalProperties": false,
                      "properties": {
                        "category": {
                          "type": "string"
                        },
                        "id": {
                          "maximum": 9007199254740991,
                          "minimum": -9007199254740991,
                          "type": "integer"
                        },
                        "label": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "category",
                        "label",
                        "updatedAt"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "id",
                    "category",
                    "label"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "user categories"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}